refactor: context 파일 분리
