+++ src/contexts/index.ts
@@ -1,0 +1,6 @@
+export {
+  useNotificationContext,
+  NotificationProvider,
+} from "./NotificationContext";
+export { useThemeContext, ThemeProvider } from "./ThemeContext";
+export { useUserContext, UserProvider } from "./UserContext";
